2675 test src folder (#3022)
- Add new test source folder menu item
- Add AbstractNewFolderWizardPagec class
