+++ org.eclipse.jdt.ui.tests/ui/org/eclipse/jdt/ui/tests/dialogs/WizardsTest.java
@@ -55,6 +55,7 @@ import org.eclipse.jdt.internal.ui.wizards.NewClassCreationWizard;
 import org.eclipse.jdt.internal.ui.wizards.NewInterfaceCreationWizard;
 import org.eclipse.jdt.internal.ui.wizards.NewPackageCreationWizard;
 import org.eclipse.jdt.internal.ui.wizards.NewSourceFolderCreationWizard;
+import org.eclipse.jdt.internal.ui.wizards.NewTestSourceFolderCreationWizard;
 
 import junit.framework.Test;
 import junit.framework.TestCase;
@@ -73,6 +74,7 @@ public class WizardsTest extends TestCase {
 		suite.addTest(new WizardsTest("testNewProjectWizard"));
 		suite.addTest(new WizardsTest("testPackageWizard"));
 		suite.addTest(new WizardsTest("testSourceFolderWizard"));
+		suite.addTest(new WizardsTest("testTestSourceFolderWizard"));
 		suite.addTest(new WizardsTest("testJarPackageWizard_sameElementOrderInJarDescriptionExports"));
 		return suite;
 	}
@@ -109,6 +111,22 @@ public class WizardsTest extends TestCase {
 		DialogCheck.assertDialog(dialog);
 
 		JavaProjectHelper.delete(jproject);
+	}
+
+	public void testTestSourceFolderWizard() throws Exception {
+		IJavaProject jproject= JavaProjectHelper.createJavaProject(PROJECT_NAME, "bin");
+		JavaProjectHelper.addSourceContainer(jproject, "src1");
+		JavaProjectHelper.addRTJar(jproject);
+
+		NewTestSourceFolderCreationWizard wizard = new NewTestSourceFolderCreationWizard();
+		wizard.init(getWorkbench(), new StructuredSelection(jproject));
+		wizard.setForcePreviousAndNextButtons(true);
+		WizardDialog dialog = new WizardDialog(getShell(), wizard);
+		dialog.create();
+		DialogCheck.assertDialog(dialog);
+
+		JavaProjectHelper.delete(jproject);
+
 	}
 
 	public void testPackageWizard() throws Exception {
