+++ org.eclipse.jdt.ui/plugin.properties
@@ -390,6 +390,9 @@ NewAnnotationType.description= Create an annotation type
 NewSourceFolderCreationWizard.label=Source Folder
 NewSourceFolderCreationWizard.description=Create a Java source folder
 
+NewTestSourceFolderCreationWizard.label=Test Source Folder
+NewTestSourceFolderCreationWizard.description=Create a Java test source folder
+
 NewJavaWorkingSetWizard.label=Java Working Set
 NewJavaWorkingSetWizard.description=Creates a Java working set
