+++ org.eclipse.jdt.ui/pom.xml
@@ -18,7 +18,7 @@
   </parent>
   <groupId>org.eclipse.jdt</groupId>
   <artifactId>org.eclipse.jdt.ui</artifactId>
-  <version>3.38.100-SNAPSHOT</version>
+  <version>3.39.0-SNAPSHOT</version>
   <packaging>eclipse-plugin</packaging>
 
 	<build>
