+++ org.eclipse.jdt.ui/ui/org/eclipse/jdt/internal/ui/actions/ActionMessages.java
@@ -437,6 +437,10 @@ public final class ActionMessages extends NLS {
 	public static String OpenNewSourceFolderWizardAction_text2;
 	public static String OpenNewSourceFolderWizardAction_description;
 	public static String OpenNewSourceFolderWizardAction_tooltip;
+	public static String OpenNewTestSourceFolderWizardAction_text;
+	public static String OpenNewTestSourceFolderWizardAction_text2;
+	public static String OpenNewTestSourceFolderWizardAction_description;
+	public static String OpenNewTestSourceFolderWizardAction_tooltip;
 	public static String GenerateBuildPathActionGroup_update_jar_text;
 	public static String GenerateBuildPathActionGroup_update_jar_description;
 	public static String GenerateBuildPathActionGroup_update_jar_tooltip;
