+++ org.eclipse.jdt.ui/ui/org/eclipse/jdt/internal/ui/actions/ActionMessages.properties
@@ -69,6 +69,7 @@ OpenTypeHierarchyAction_messages_no_java_elements=One or more Java elements must
 OpenNewAnnotationWizardAction_description=Opens the new annotation type wizard
 OpenNewJavaProjectWizardAction_description=Opens the new Java project wizard
 OpenNewSourceFolderWizardAction_description=Opens the new Java source folder wizard
+OpenNewTestSourceFolderWizardAction_description=Opens the new Java source folder wizard
 OpenNewInterfaceWizardAction_description=Opens the new interface wizard
 OpenTypeHierarchyAction_messages_no_valid_java_element=Either a single type or member, or one or more packages, source folders, etc. must be selected.
 
@@ -326,10 +327,13 @@ OpenNewJavaProjectWizardAction_tooltip=Opens the new Java project wizard
 OpenNewPackageWizardAction_description=Opens the new Java package wizard
 OpenViewActionGroup_showInAction_label=Sho&w In
 OpenNewSourceFolderWizardAction_tooltip=Opens the new Java source folder wizard
+OpenNewTestSourceFolderWizardAction_tooltip=Opens the new Java test source folder wizard
 OpenNewClassWizardAction_description=Opens the new class wizard
 OpenNewInterfaceWizardAction_tooltip=Opens the new interface wizard
 OpenNewSourceFolderWizardAction_text=Source Folder
 OpenNewSourceFolderWizardAction_text2=&New Source Folder...
+OpenNewTestSourceFolderWizardAction_text=Test Source Folder
+OpenNewTestSourceFolderWizardAction_text2=New Test Source Folder...
 OpenNewEnumWizardAction_description=Opens the new enum type wizard
 OpenNewRecordWizardAction_description=Opens the new record type wizard
 OpenNewJavaProjectWizardAction_text=Java Project
