+++ org.eclipse.jdt.ui/ui/org/eclipse/jdt/internal/ui/wizards/NewTestSourceFolderCreationWizard.java
@@ -1,0 +1,66 @@
+/*******************************************************************************
+ * Copyright (c) 2000, 2011 IBM Corporation and others.
+ *
+ * This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License 2.0
+ * which accompanies this distribution, and is available at
+ * https://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *     IBM Corporation - initial API and implementation
+ *******************************************************************************/
+package org.eclipse.jdt.internal.ui.wizards;
+
+import org.eclipse.core.runtime.CoreException;
+import org.eclipse.core.runtime.IProgressMonitor;
+
+import org.eclipse.jdt.core.IJavaElement;
+
+import org.eclipse.jdt.internal.ui.JavaPlugin;
+import org.eclipse.jdt.internal.ui.JavaPluginImages;
+
+
+public class NewTestSourceFolderCreationWizard extends NewElementWizard {
+
+	private NewTestSourceFolderWizardPage fPage;
+
+	public NewTestSourceFolderCreationWizard() {
+		super();
+		setDefaultPageImageDescriptor(JavaPluginImages.DESC_WIZBAN_NEWSRCFOLDR);
+		setDialogSettings(JavaPlugin.getDefault().getDialogSettings());
+		setWindowTitle(NewWizardMessages.NewTestSourceFolderWizardPage_title);
+	}
+
+	/*
+	 * @see Wizard#addPages
+	 */
+	@Override
+	public void addPages() {
+		super.addPages();
+		fPage= new NewTestSourceFolderWizardPage();
+		addPage(fPage);
+		fPage.init(getSelection());
+	}
+
+	@Override
+	protected void finishPage(IProgressMonitor monitor) throws InterruptedException, CoreException {
+		fPage.createPackageFragmentRoot(monitor); // use the full progress monitor
+	}
+
+	@Override
+	public boolean performFinish() {
+		boolean res= super.performFinish();
+		if (res) {
+			selectAndReveal(fPage.getCorrespondingResource());
+		}
+		return res;
+	}
+
+	@Override
+	public IJavaElement getCreatedElement() {
+		return fPage.getNewPackageFragmentRoot();
+	}
+
+}
