+++ org.eclipse.jdt.ui/ui/org/eclipse/jdt/internal/ui/wizards/NewWizardMessages.java
@@ -321,6 +321,10 @@ public final class NewWizardMessages extends NLS {
 	public static String NewSourceFolderWizardPage_ReplaceExistingSourceFolder_label;
 	public static String NewSourceFolderWizardPage_edit_description;
 
+	public static String NewTestSourceFolderWizardPage_title;
+	public static String NewTestSourceFolderWizardPage_description;
+	public static String NewTestSourceFolderWizardPage_edit_description;
+
 	public static String BuildPathsBlock_tab_source;
 	public static String BuildPathsBlock_tab_projects;
 	public static String BuildPathsBlock_tab_libraries;
