+++ org.eclipse.jdt.ui/ui/org/eclipse/jdt/internal/ui/wizards/NewWizardMessages.properties
@@ -300,6 +300,10 @@ NewSourceFolderWizardPage_warning_ReplaceSF=To avoid overlapping, the existing p
 NewSourceFolderWizardPage_warning_AddedExclusions_singular=Exclusion patterns of source folder ''{0}'' updated to solve nesting.
 NewSourceFolderWizardPage_warning_AddedExclusions_plural=Exclusion patterns of {0} source folders updated to solve nesting.
 
+NewTestSourceFolderWizardPage_title=Test Source Folder
+NewTestSourceFolderWizardPage_description=Create a new test source folder.
+NewTestSourceFolderWizardPage_edit_description=Edit source folder
+
 # ------- BuildPathsBlock -------
 
 BuildPathsBlock_tab_source=&Source
